+++ snippy/meta.py
@@ -25,7 +25,7 @@ __email__ = 'laaksonen.heikki.j@gmail.com'
 __homepage__ = 'https://github.com/heilaaks/snippy'
 __license__ = 'GNU Affero General Public License v3 or later'
 __title__ = 'snippy'
-__version__ = '0.10a0'
+__version__ = '0.10.0'
 
 __docs__ = 'http://snippy.readthedocs.io/en/v' + __version__ + '/'
 __openapi__ = 'https://app.swaggerhub.com/apis/heilaaks/snippy/' + __version__