+++ CHANGELOG.rst
@@ -38,7 +38,8 @@ Security
 ~~~~~~~~
 
 * Remove setuid/setgid bit from binaries in Docker image.
-* Fix server bind IP address in Docker container.
+* Remove unnecessary file permissions in Docker image.
+* Change IP address where server binds in Docker container.
 
 0.9.0 (2019-02-02)
 ------------------